requirements.txt
torch
torchaudio
gradio>=4.0.0
coqui-tts
soundfile
numpy
